+++ tests/uniswap_v2/sendSwap.ts
@@ -60,7 +60,8 @@ const swapUniswapRawSender = async (amountsIn: number[], amountOutMin: number, t
     for (const amount of amountsIn) {
         assets.push({
             address: tokenAddress,
-            amount: amount,
+            amountWithoutDecimals: amount,
+            decimals: 9,
         });
     }